backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
requests
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]